"auth.ts moved into lib folder"
