+++ src/lib/auth.ts
